Revert "Refactor TypicalPersons and TypicalJobs to TypicalObjects"
This reverts commit 5308f431b08fe19aee0ee098a05d7c0ba22a7272.
