+++ src/test/java/seedu/address/model/AddressBookTest.java
@@ -5,9 +5,9 @@ import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertTrue;
 import static seedu.address.logic.commands.CommandTestUtil.VALID_ADDRESS_BOB;
 import static seedu.address.logic.commands.CommandTestUtil.VALID_TAG_HUSBAND;
-import static seedu.address.testutil.TypicalObject.ALICE;
-import static seedu.address.testutil.TypicalObject.ENGINEER;
-import static seedu.address.testutil.TypicalObject.getTypicalAddressBook;
+import static seedu.address.testutil.TypicalJobs.ENGINEER;
+import static seedu.address.testutil.TypicalPersons.ALICE;
+import static seedu.address.testutil.TypicalPersons.getTypicalAddressBook;
 
 import java.util.Arrays;
 import java.util.Collection;
