+++ src/test/java/seedu/address/ui/BrowserPanelTest.java
@@ -2,7 +2,7 @@ package seedu.address.ui;
 
 import static guitests.guihandles.WebViewUtil.waitUntilBrowserLoaded;
 import static org.junit.Assert.assertEquals;
-import static seedu.address.testutil.TypicalObject.ALICE;
+import static seedu.address.testutil.TypicalPersons.ALICE;
 
 import java.net.URL;
